requirements.txt
Django==4.0.4
python-dotenv==0.21.0
whitenoise==6.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
django-heroku==0.3.1
dj-database-url==1.2.0
requests==2.28.1
pytz==2022.7.1